refactor: ProjectTasksView in declarative view code
