+++ Pharus/Components/Project Tasks/ProjectTasksDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  ProjectTasksDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 28/07/22.
+//
+
+import Foundation
+
+protocol ProjectTaskDelegate: AnyObject {
+    func checkmarkButtonTapped(task: TaskModel)
+}
